+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Tous les changements de ce projet seront documentés dans ce fichier.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## 24-11-2021
+
+- Ajout de la structure de la base de donnée en script SQL dans assets/sql/php_project.sql
+
 ## 22-11-2021
 
 ### Added